% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{basis}
\alias{basis}
\title{Basis Functions}
\usage{
basis(x, ...)
}
\arguments{
\item{x}{Predictor vector.}

\item{scale}{Scale parameter (e.g. width of the interval over which the basis function is defined.).}

\item{skewness}{Skewness parameter of basis function. Zero value implies no skewness, negative values are
left-skewed and positive values are right-skewed.}

\item{degree}{Degree of polynomial basis.}
}
\description{
Basis function definitions.
}
