% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass.grid}
\alias{biomass.grid}
\title{Calculate Biomass}
\usage{
biomass.grid(year, category = "COM")
}
\arguments{
\item{year}{Survey year.}

\item{category}{Snow crab category.}

\item{weight}{Logical value specifying whether to convert catches to weights.}
}
\description{
Functions to calculate snow crab biomass.
}
\examples{
r <- biomass.grid(2022)
r <- biomass.grid(2022, category = "MMSC12") # Fishery recruits.
}
