% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{djohnson}
\alias{djohnson}
\alias{pjohnson}
\alias{rjohnson}
\title{Probability Distributions}
\usage{
djohnson(x, eta = 0, lambda = 1, delta = 1, gamma = 0, log = FALSE)

pjohnson(x, eta = 0, lambda = 1, delta = 1, gamma = 0, log = FALSE)

rjohnson(n, eta = 0, lambda = 1, delta = 1, gamma = 0)
}
\arguments{
\item{x}{Domain values.}

\item{eta}{Location parameter.}

\item{lambda}{Scale parameter.}

\item{delta}{Shape parameter.}

\item{gamma}{Skewness parameter parameter.}

\item{log}{Logical value specifying whether to return the logarithmic values of the function.}

\item{n}{Number of random variates to generate.}
}
\description{
Set of probability distributions.
}
\section{Functions}{
\itemize{
\item \code{djohnson()}: Probability density function for the Johnson's SU distribution.

\item \code{pjohnson()}: Cumulative density function for the Johnson's SU distribution.

\item \code{rjohnson()}: Cumulative density function for the Johnson's SU distribution.

}}
