% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.minilog.R
\name{fit.minilog}
\alias{fit.minilog}
\title{Initial Touchdown and Lift-Off Times}
\usage{
\method{fit}{minilog}(x, truncate = FALSE)
}
\arguments{
\item{x}{Minilog data object.}

\item{truncate}{Touch}
}
\description{
Functions to find initial parameter values for Minilog data.
}
\examples{
x <- read.minilog(year = 2014, tow.id = tow.id, location = "tilt")
}
