% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit.variogram}
\alias{fit.morphometry.scsbio}
\title{Fit Statistical Models}
\usage{
fit(x, ...)

\method{fit}{variogram}(x, model = "spherical", nugget = TRUE, distance.exponent = 0, inits, ...)

\method{fit}{morphometry.scsbio}(x, y, z, n = 1, sex, theta, discrete = FALSE, print = FALSE, trace = 0)
}
\arguments{
\item{x}{Data object or numeric vector representing a morphometric predictor variable..}

\item{model}{Character string specifying the model type.}

\item{nugget}{Logical value specifying whether a variogram model contains a nugget semi-variance parameter.}

\item{distance.exponent}{Numeric value specifying the exponent to be applied in the distance metric.

export}

\item{y}{Numeric vector representing a morphometric response variable.}

\item{z}{Binary classification vector classifying morphometric data into discrete maturity groups.}

\item{n}{Number of observations for each (x,y) pair (optional).}

\item{sex}{Biological sex which specifies the type of analysis or initial values to be applied.}

\item{theta}{Parameter vector.}

\item{discrete}{Logical value specifying whether to treat observations as discrete rounded values.}

\item{print}{Logical value specifying whether to print information about progress during model fitting.}

\item{trace}{Optimization control output (see \code{\link[stats]{optim}}).}
}
\description{
Functions to fit statistical models to data.
}
\section{Methods (by class)}{
\itemize{
\item \code{fit(variogram)}: Fit a model to empirical variogram data.

\item \code{fit(morphometry.scsbio)}: Fit a morphometric model to snow crab morphometric data.

}}
