% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.star.oddi.R
\name{init}
\alias{init}
\title{Initial Touchdown and Lift-Off Times}
\usage{
init(x, ...)
}
\arguments{
\item{x}{Star Oddi data object.}

\item{truncate}{Touch}
}
\description{
Functions to find initial parameter values for Star Oddi data.
}
\examples{
x <- read.star.oddi(year = 2014, tow.id = tow.id, location = "tilt")
}
