% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.times.star.oddi.R, R/init.star.oddi.R
\name{init.tilt}
\alias{init.tilt}
\title{Estimate Trawl Event Times}
\usage{
\method{init}{tilt}(x, init)

\method{init}{tilt}(x, init)
}
\arguments{
\item{x}{Star Oddi data object.}

\item{t}{Time vector.}

\item{theta}{Named free parameter vector.}

\item{fixed}{Named fixed parameter vector.}

\item{model}{Model to be used for modelling bottom trawling phase.}

\item{y}{Pressure observations.
s}

\item{truncate}{Touch}
}
\description{
Functions to find initial parameter values for Star Oddi data.

Functions to find initial parameter values for Star Oddi data.
}
\examples{
x <- read.star.oddi(year = 2014, tow.id = tow.id, location = "tilt")
x <- read.star.oddi(year = 2014, tow.id = tow.id, location = "tilt")
}
