% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Initialize Model Parameters}
\usage{
initialize(x, ...)
}
\arguments{
\item{x, y}{Data vectors.

Initialize PLM parameters}
}
\description{
Methods to provide initial parameter estimates for a given statistical model.
}
