% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{ked}
\alias{ked}
\alias{ked.default}
\alias{ked.scsset}
\title{Kriging Spatial Modelling}
\usage{
ked(x)

\method{ked}{default}(x, x0, y, z, z0, nugget = 0, sill, range, positive = TRUE)

\method{ked}{scsset}(
  x,
  y,
  variables,
  variogram,
  year,
  category,
  weight = FALSE,
  hard.shelled = FALSE,
  lag = 3,
  max.distance = 75,
  variogram.average = 1,
  xlim = c(0, 440),
  ylim = c(0, 400),
  grid = c(100, 100),
  long.ref = -66,
  lat.ref = 45.5,
  n = 8,
  nugget,
  sill,
  range,
  polygon = kriging.polygons()$gulf,
  bug = FALSE,
  ...
)
}
\arguments{
\item{x}{Data coordinates or a data frame containing coordinates.}

\item{x0}{Prediction coordinates.}

\item{y}{Data response values.}

\item{z}{External drift values at data locations.}

\item{z0}{External drift values at prediction locations.}

\item{nugget}{Numerical value specifying the nugget value for the variogram model.}

\item{sill}{Numerical value specifying the sill or asymptotic value for the variogram model.}

\item{range}{Numerical value specifying the range value for the variogram model.}

\item{positive}{Logical value specifying whether to constrain kriged interpolation to be positive values.}

\item{variables}{Character string specifying the varigram in \code{x} to be analyzed.}

\item{variogram}{\code{variogram} object.}

\item{year}{Snow crab survey year.}

\item{category}{Snow crab survey year.}

\item{weight}{= FALSE,}

\item{hard.shelled}{= FALSE,}

\item{lag}{Lag distance, in kilometers, used for calculating the empirical variogram (i.e. bin size).}

\item{max.distance}{Maximum distance used for calculating the empirical variogram.}

\item{variogram.average}{Number of years to use in the variogram averaging.}

\item{xlim, ylim}{Horizontal and vertical limits in kilometers with respect to the reference coordinates (\code{long.ref} and \code{lat.ref}).}

\item{grid}{Two-element vector specifying the dimensions of the interpolation prediction grid.}

\item{long.ref, lat.ref}{Lower-left coordinates used, along with \code{grid}, to define the interpolation grid over which kriging will be performed.}

\item{n}{Number of nearest neighbours used when kriging}

\item{polygon}{Polygon coordinates for within which kriging is to be performed.}

\item{bug}{Logical value specifying whether to include a bug in the calculations which was present in past analyses.}

\item{nugget, sill, range}{Variogram model parameter values.}
}
\description{
Spatial interpolation and integration funcions using kriging.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Perform kriging with external drift.

\item \code{scsset}: Perform kriging with external drift for snow crab data.
}}

