% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.intersect.R
\name{line.intersect}
\alias{line.intersect}
\title{Line intersections}
\usage{
line.intersect(beta0, beta1, pairwise = TRUE)
}
\arguments{
\item{beta0}{Two-element vector or two-column matrix of line slopes and intercept values.}

\item{beta1}{Two-element vector or two-column matrix of line slopes and intercept values.}

\item{pairwise}{Logical value specifying whether to calculate intersection points for all possible slope and intercept pair combinations.}
}
\value{
Returns a list of x and y intersection points.
}
\description{
Returns coordinates where pairs of lines intersect.
}
\examples{
n <- 10
m <- 2*runif(n)-1
b <- 2*runif(n)-1
plot(c(-2, 2), c(-2, 2), type = "n", xlab = "x", ylab = "y")
for (i in 1:n) abline(b[i], m[i], col = "grey70")
r <- line.intersect(cbind(m, b))
points(r$x, r$y, pch = 21, bg = "red")
}
