% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.likelihood.R
\name{loglike}
\alias{loglike}
\title{Model Log-likelihood functions.}
\usage{
loglike(x, ...)
}
\arguments{
\item{x, y}{Data vectors.}

\item{theta}{Parameter vector.}

\item{fixed}{Vector of fixed parameter values.}

\item{precision}{Precision of observed data values.}
}
\description{
Methods for calculating log-likelihood values for various types of models.
}
