% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphometric.maturity.R
\name{morphometric.maturity}
\alias{morphometric.maturity}
\alias{morphometric.maturity.scsbio}
\title{Determine Morphometric Maturity}
\usage{
morphometric.maturity(x, ...)

\method{morphometric.maturity}{scsbio}(x, probability = FALSE, ...)
}
\arguments{
\item{x}{Data object.}

\item{probability}{Logical value specifying whether the morphoemtric data are rounded.}
}
\description{
Functions to determine sexual maturity from morphometric measurements.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsbio}: Determine morphometric maturity from snow crab biological data.
}}

\examples{
x <- read.scsbio(2010)
v <- morphometric.maturity(x)

}
\seealso{
\code{\link{morphometry}}

\code{\link[gulf.data]{maturity}}
}
