% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphometry.R
\name{morphometry}
\alias{morphometry}
\alias{morphometry.default}
\alias{morphometry.scsbio}
\title{Analyze Morphometric Data}
\usage{
morphometry(x, ...)

\method{morphometry}{default}(x, y, species, kurtotic = FALSE, ...)

\method{morphometry}{scsbio}(x, y, z, theta, sex, fit = TRUE, discrete = FALSE)
}
\arguments{
\item{x}{Predictor variable.}

\item{y}{Response variable.}

\item{z}{Binary classification variable (optional).}

\item{theta}{Named model parameter vector.}

\item{sex}{Biological sex.}

\item{fit}{Logical value specifying whether to fit the model to the data.}

\item{discrete}{Logical value specifying whether the morphometric data are rounded.}
}
\description{
Functions to analyze morphometric data.
}
\section{Methods (by class)}{
\itemize{
\item \code{morphometry(default)}: Perform morphometric regression analysis.

\item \code{morphometry(scsbio)}: Perform morphometric regression analysis for snow crab biological data.

}}
\examples{
x <- read.scsbio(2010)
v <- morphometry(x)
}
