% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.R
\name{plm.basis}
\alias{plm.basis}
\title{Piecewise-Linear Model Class}
\usage{
plm.basis(x, scale = 1, skewness = 0, degree = 3, ...)
}
\arguments{
\item{x}{Numerical vector of values at which the function is to be evaluated.}

\item{scale}{Numeric value specifying the total width of the range of the basis functions.}

\item{skewness}{Degree of asymmetry in the basis functions. Negative values yield left-skewed basis functions,
right values yield right-skewed basis functions, while a zero value indicates no skewness.}

\item{degree}{Integer specifying the degree of the polynomial basis function.}
}
\description{
Set of functions which define a class of smooth piecewise linear models.
}
\examples{


}
