% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.variogram}
\alias{plot.variogram}
\alias{plot.morphometry.scsbio}
\title{Plot Statistical Models}
\usage{
\method{plot}{variogram}(
  x,
  scale,
  cex = 1.5,
  cex.axis = 1.25,
  xaxt = "s",
  yaxt = "s",
  xlab,
  ylab,
  xlim,
  ylim,
  nugget = TRUE,
  sill = TRUE,
  range = TRUE,
  ...
)

\method{plot}{morphometry.scsbio}(x, y, theta, xlim = c(10, 140), log = TRUE, title, discrete = FALSE, ...)
}
\arguments{
\item{x}{\code{variogram} object.}

\item{scale}{Scalar by which to scale the variogram y axis.}

\item{cex}{Text size.}

\item{cex.axis}{Text size for plot axes.}

\item{xaxt, yaxt}{\code{\link{par}} graphics parameters.}

\item{xlab, ylab}{Character strings specifying the axes labels.}

\item{xlim, ylim}{Two-element numerical vectors specifying plot axes limits.}

\item{nugget, sill, range}{Logical value specifying whether to display reference lines for variogram parameters.}

\item{discrete}{Logical value specifying whether morphometric data are rounded.}
}
\description{
Functions to graphically display statistical models.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot(variogram)}: Variogram plot method.

\item \code{plot(morphometry.scsbio)}: Morphometric plot for snow crab biological data.

}}
