% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R, R/integrate.R, R/polynomial.R
\name{derive.polynomial}
\alias{derive.polynomial}
\alias{derive.spline}
\alias{integrate.polynomial}
\alias{integrate.spline}
\alias{polynomial}
\alias{polynomial.default}
\alias{+.polynomial}
\alias{-.polynomial}
\alias{*.polynomial}
\alias{print.polynomial}
\alias{coef.polynomial}
\title{Polynomial Class}
\usage{
\method{derive}{polynomial}(x)

\method{derive}{spline}(x)

\method{integrate}{polynomial}(x)

\method{integrate}{spline}(x, n = 1)

polynomial(beta)

\method{polynomial}{default}(beta)

\method{+}{polynomial}(x, y)

\method{-}{polynomial}(x, y)

\method{*}{polynomial}(x, y)

\method{print}{polynomial}(x)

\method{coef}{polynomial}(x)
}
\arguments{
\item{x, y}{Polynomial object.}

\item{beta}{Numeric vector of polynomial coefficients. First element is the constant coefficient,
the second element is the linear coefficient, the third is the quadratic coefficient, etc...}
}
\description{
Methods for creating, evaluating, manupulating and displaying polynomial functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{polynomial(default)}: Polynomial object creation and evaluation.

}}
\section{Methods (by generic)}{
\itemize{
\item \code{derive(polynomial)}: Polynomial derivation method.

\item \code{integrate(polynomial)}: Polynomial integration method.

\item \code{ + }: Polynomial unary and binary addition operator.

\item \code{ - }: Polynomial unary and binary substraction operator.

\item \code{ * }: Polynomial multiplication operator.

\item \code{print(polynomial)}: Polynomial print method.

\item \code{coef(polynomial)}: Polynomial print method.

}}
\section{Functions}{
\itemize{
\item \code{derive(spline)}: Spline derivation method.

\item \code{integrate(spline)}: Spline integration method.

}}
\examples{
p1 <- polynomial(1:3)    # Create polynomial 1 + 2*x + 3*x^2.
p2 <- polynomial(0:5)    # Create polynomial x + 2*x^2 + 3*x^3 + 4*x^4 + 5*x^5.

# Evaluation:
x <- 1:10
polynomial(1)(x)        # Constant function.
polynomial(c(1,0,2))(x) # Quadratic 1 + 2*x^2.
p1(x)                   # Evaluate 'p1' at 'x'.
(p1 + p2)(x)            # Evaluate the sum of 'p1' and 'p2' at 'x'.
(p1 - p2)(x)            # Evaluate the difference of 'p1' and 'p2' at 'x'.
}
