% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{polyval}
\alias{polyval}
\title{Polynomial evaluation}
\usage{
polyval(x, p)
}
\arguments{
\item{x}{Numeric vector at which the polynomial will be evaluated.}

\item{p}{Polynomial coefficients. The first coefficient corresponds to the intercept 
or constant coefficent, while the last element corresponds to the coefficient
of the highest polynomial order.}
}
\description{
Horner's method for polynomial evaluation.
}
\examples{
x <- 1:10
polyval(x, 1) # Constant function.
polyval(x, c(1,0,2)) # Quadratic 1 + 2*x^2.
}
