% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma}
\alias{sigma}
\title{Calculate Error}
\usage{
sigma(x, ...)
}
\arguments{
\item{x}{Predictor vector.}

\item{theta}{Parameter vector.}
}
\description{
Calculate regression error for various statistical models.
}
