% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical value specifying whether to remove NA values from the calculation.}
}
\description{
Calculate sample skewness.
}
\examples{
x <- exp(rnorm(100))
skewness(x)
}
