% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{spline}
\alias{spline}
\title{Spline Functions}
\usage{
spline(x)
}
\arguments{
\item{polynomial}{Polynomial coefficients, polynomial object or list of polynomial objects.}

\item{knots}{Numeric vector of spline knots.}
}
\description{
Spline function class
}
