% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ked}
\alias{summary.ked}
\title{Summary for Statistical Models}
\usage{
\method{summary}{ked}(x, polygon, ...)
}
\description{
Generate a summary for a statistical model.
}
\section{Methods (by class)}{
\itemize{
\item \code{summary(ked)}: Generate a summary of a 'ked' object. The output includes sample and kriged estimates,
along with error estimates.

}}
