% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{parameters.cpp}
\alias{parameters.cpp}
\alias{data.cpp}
\alias{update.parameters}
\alias{update.map}
\title{Template Model Builder (TMB) Functions}
\usage{
parameters.cpp(x)

data.cpp(x)

\method{update}{parameters}(x, obj, fixed, random, map)

\method{update}{map}(map, free.variables, fixed.variables)
}
\description{
Functions for use with Template Model Builder (TMB).
}
\section{Functions}{
\itemize{
\item \code{parameters.cpp()}: Extract model parameter name(s) from TMB model.

\item \code{data.cpp()}: Extract data variable name(s) from TMB model.

\item \code{update(parameters)}: Update model parameters using TMB model output.

\item \code{update(map)}: Update model parameter map.

}}
