% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/variogram.R
\name{fit.variogram}
\alias{fit.variogram}
\alias{variogram}
\alias{variogram.default}
\alias{variogram.scsset}
\alias{variogram.variogram}
\title{Empirical Variograms}
\usage{
\method{fit}{variogram}(x, model = "spherical", nugget = TRUE, distance.exponent = 0, inits, ...)

variogram(x, ...)

\method{variogram}{default}(
  x,
  z,
  lag,
  max.distance,
  mean.lag = TRUE,
  distance.units,
  response.units,
  trim,
  fit = TRUE,
  ...
)

\method{variogram}{scsset}(
  x,
  category,
  variable,
  weight = FALSE,
  lag = 3,
  max.distance = 75,
  distance.exponent = 1,
  ...
)

\method{variogram}{variogram}(x, h)
}
\arguments{
\item{x}{Data observations, coordinates or a distance matrix.}

\item{model}{Character string specifying a variogram function type.}

\item{nugget}{Logical value specifying whether the variogram model includes a nugget (i.e. zero-lag) variance component.}

\item{distance.exponent}{Positive numeric value specifying the exponent of the distance metric.}

\item{inits}{Initial parameter values.}

\item{z}{Vector of data observations.}

\item{h}{Lag distances.}
}
\description{
Functions to calculate empirical variograms and fit variogram models to data.
}
\section{Methods (by class)}{
\itemize{
\item \code{variogram}: Fit a model to empirical variogram data.

\item \code{default}: Calculate empirical variogram.

\item \code{scsset}: Calculate empirical variogram for snow crab survey data.

\item \code{variogram}: Evaluate variogram at specified distances.
}}

