% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphometric.maturity.R
\name{morphometric.maturity}
\alias{morphometric.maturity}
\alias{morphometric.maturity.scsbio}
\title{Determine Morphometric Maturity}
\usage{
morphometric.maturity(x, ...)

\method{morphometric.maturity}{scsbio}(x, probability = FALSE, ...)
}
\arguments{
\item{x}{Data object.}

\item{probability}{Logical value specifying whether the morphometric data are rounded.}
}
\description{
Functions to determine sexual maturity from morphometric measurements.
}
\section{Methods (by class)}{
\itemize{
\item \code{morphometric.maturity(scsbio)}: Determine morphometric maturity from snow crab biological data.

}}
\examples{
# Male morphometric maturity:
x <- read.scsbio(2011, sex = 1)
v <- morphometric.maturity.scsbio(x)
plot(c(0, 140), c(0, 40), type = "n", xlab = "Carapace width(m)", ylab = "Chela height(m)", xaxs = "i", yaxs = "i")
grid()
points(x$carapace.width[v], x$chela.height[v], pch = 21, bg = "red", cex = 0.2, lwd = 0.2)
points(x$carapace.width[!v], x$chela.height[!v], pch = 21, bg = "blue", cex = 0.2, lwd = 0.2)
box()

}
\seealso{
\code{\link{morphometry}}

\code{\link[gulf.data]{maturity}}
}
