% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram.R
\name{variogram}
\alias{variogram}
\alias{variogram.default}
\alias{variogram.scsset}
\alias{variogram.variogram}
\title{Empirical Variograms}
\usage{
variogram(x, ...)

\method{variogram}{default}(
  x,
  z,
  lag,
  max.distance,
  mean.lag = TRUE,
  distance.units,
  response.units,
  trim,
  fit = TRUE,
  ...
)

\method{variogram}{scsset}(
  x,
  category,
  variable,
  weight = FALSE,
  lag = 3,
  max.distance = 75,
  distance.exponent = 1,
  ...
)

\method{variogram}{variogram}(x, h)
}
\arguments{
\item{x}{Data observations, coordinates or a distance matrix.}

\item{z}{Vector of data observations.}

\item{distance.exponent}{Positive numeric value specifying the exponent of the distance metric.}

\item{h}{Lag distances.}

\item{model}{Character string specifying a variogram function type.}

\item{nugget}{Logical value specifying whether the variogram model includes a nugget (i.e. zero-lag) variance component.}

\item{inits}{Initial parameter values.}
}
\description{
Functions to calculate empirical variograms and fit variogram models to data.
}
\section{Methods (by class)}{
\itemize{
\item \code{variogram(default)}: Calculate empirical variogram.

\item \code{variogram(scsset)}: Calculate empirical variogram for snow crab survey data.

\item \code{variogram(variogram)}: Evaluate variogram at specified distances.

}}
